Frontend refactor: create 'third_party' module (#720)
Move external projects like OQC support and CUDA compiler support to a
new module `catalyst/third_party`.

It's unclear whether `third_party` should be exposed as is, or the
contents be re-exported to catalyst (e.g. to access `catalyst.cuda`).

[sc-59509]

---------
