+++ MANIFEST.in
@@ -2,4 +2,4 @@ recursive-include frontend/catalyst/bin *
 recursive-include frontend/catalyst/lib *
 recursive-include frontend/catalyst/enzyme *
 recursive-include frontend/mlir_quantum *
-recursive-include frontend/catalyst/cuda/ *.toml
+recursive-include frontend/catalyst/third_party/cuda/ *.toml
