+++ doc/code/__init__.rst
@@ -20,10 +20,10 @@ Module: catalyst.debug
     :no-inheritance-diagram:
     :fullname:
 
-Module: catalyst.cuda
----------------------
+Module: catalyst.third_party.cuda
+---------------------------------
 
-.. automodapi:: catalyst.cuda
+.. automodapi:: catalyst.third_party.cuda
     :no-heading:
     :no-inheritance-diagram:
     :fullname:
