+++ frontend/catalyst/utils/runtime.py
@@ -44,7 +44,7 @@ DEFAULT_LIB_PATHS = {
     "llvm": os.path.join(package_root, "../../../mlir/llvm-project/build/lib"),
     "runtime": os.path.join(package_root, "../../../runtime/build/lib"),
     "enzyme": os.path.join(package_root, "../../../mlir/Enzyme/build/Enzyme"),
-    "oqc_runtime": os.path.join(package_root, "../../catalyst/oqc/src/build"),
+    "oqc_runtime": os.path.join(package_root, "../../catalyst/third_party/oqc/src/build"),
 }
