+++ runtime/Makefile
@@ -1,4 +1,4 @@
-PYTHON := $(shell which python3)
+PYTHON?=$(shell which python3)
 C_COMPILER?=$(shell which clang)
 CXX_COMPILER?=$(shell which clang++)
 COMPILER_LAUNCHER?=$(shell which ccache)
